requirement.txt
shapely==1.8.5.post1
lyft_dataset_sdk
numba==0.48.0
plyfile
matplotlib==3.4.0
networkx==2.2
nuscenes-devkit==1.1.6
pandas==1.4.4
Pillow==8.3.2
scikit-image==0.19.3
scikit-learn==1.3.1
scipy==1.10.1
urllib3==2.0.5
numpy==1.19.5
openai